Remove NumberHelper monkeypatching and alias_method_chain
